scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "mapbox-vector-tile>=2.2.0",
    "mercantile>=1.2.1",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
